Fix type check
